+++ tango-pvp/src/game/bn6.rs
@@ -285,7 +285,7 @@ impl crate::hooks::Hooks for Hooks {
                     let match_ = match_.clone();
                     let munger = self.munger();
                     Box::new(move |mut core| {
-                        let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                        let pc = core.as_ref().gba().cpu().thumb_pc();
                         core.gba_mut().cpu_mut().set_thumb_pc(pc + 6);
                         munger.set_copy_data_input_state(core, if match_.blocking_lock().is_some() { 2 } else { 4 });
                     })
@@ -517,7 +517,7 @@ impl crate::hooks::Hooks for Hooks {
                 {
                     let munger = self.munger();
                     Box::new(move |mut core| {
-                        let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                        let pc = core.as_ref().gba().cpu().thumb_pc();
                         core.gba_mut().cpu_mut().set_thumb_pc(pc + 6);
                         munger.set_copy_data_input_state(core, 2);
                     })
@@ -730,7 +730,7 @@ impl crate::hooks::Hooks for Hooks {
                     if !stepper_state.disable_bgm() {
                         return;
                     }
-                    let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                    let pc = core.as_ref().gba().cpu().thumb_pc();
                     core.gba_mut().cpu_mut().set_thumb_pc(pc + 4);
                 })
             }),
@@ -759,7 +759,7 @@ impl crate::hooks::Hooks for Hooks {
                 {
                     let munger = self.munger();
                     Box::new(move |mut core| {
-                        let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                        let pc = core.as_ref().gba().cpu().thumb_pc();
                         core.gba_mut().cpu_mut().set_thumb_pc(pc + 6);
                         munger.set_copy_data_input_state(core, 2);
                     })
